requirements.txt
# Core framework and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication and Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6

# MinIO Object Storage
minio==7.2.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1
email-validator==2.1.0

# Environment variable management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# HTTP client for external calls
httpx==0.25.2

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Additional utilities
python-dateutil==2.8.2
passlib==1.7.4
